requirements.txt
anyio==3.6.1
asgiref==3.5.2
asyncer==0.0.1
atomicwrites==1.4.0
attrs==21.4.0
certifi==2020.6.20
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
contourpy==1.0.7
cycler==0.11.0
docopt==0.6.2
fastapi==0.78.0
filetype==1.1.0
fonttools==4.33.3
gdown==4.5.1
h11==0.13.0
idna==3.3
ImageHash==4.2.1
imageio==2.19.3
iniconfig==1.1.1
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.6.3
networkx==3.0
numpy==1.22.4
opencv-python-headless==4.7.0.68
packaging==21.3
Pillow==9.1.1
pipreqs==0.4.13
pluggy==1.0.0
py==1.11.0
pydantic==1.9.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-multipart==0.0.5
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
requests==2.27.1
scikit-image==0.18.3
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
threadpoolctl==3.1.0
tifffile==2022.5.4
tomli==2.0.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0+cpu
--find-links https://download.pytorch.org/whl/torch_stable.html
torchvision==0.12.0+cpu
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
wincertstore==0.2
yarg==0.1.9
